Use possible owner's full name in emails
The old custom user model that was removed in 5d20475 returned the
user's full name (albeit as "Last, First") from its `__str__` method.

We now use the approach of passing the user to the template where it's
used and then doing `{{ user.get_full_name }}` (instead of passing
`str(user)` to the template).