Fix some tests that were "broken" by the upgrade to Python 3.5
It appears that these tests were already broken in that they weren't
actually testing anything. It seems there was some change in the mock
package that brought this to light (it seems to be stricter now?).

The fix is to not use mock for these tests since it's not required nor
is it useful.

There are several other tests that use mock; we should probably look at
those too and make sure they're actually doing something useful.